log manager handles devices correctly
